// src/app/modules/main/pages/main-rooms/models/room.model.ts
export interface IRoomModel {
  moderatorFullname: string;
  updateDate: string;
  createUserId: string;
  updateUserId: string;
  conferenceTypeNameAz: string;
  moderator: string;
  active: string;
  joinStatus: string;
  userId: string;
  conferenceTypeName: string;
  conferenceType: string;
  moderatorEmail: string;
  moderatorStatus: string;
  conferenceTypeNameEn: string;
  name: string;
  userFullname: string;
  userEmail: string;
  id: string;
  conferenceTypeNameRu: string;
  createDate: string;
}